requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
filelock==3.3.1
idna==3.3
platformdirs==2.4.0
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
virtualenv==20.9.0
